% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mostprobable.R
\name{mostprobable}
\alias{mostprobable}
\title{This function calculates the moste probable playable row given a table of odds for thirteen games}
\usage{
mostprobable(table = NONE)
}
\arguments{
\item{table}{of odds, default = NONE}
}
\description{
This function calculates the moste probable playable row given a table of odds for thirteen games
}
\examples{
mostprobable()
}
\keyword{cats}

