% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilites.R
\name{probabilities}
\alias{probabilities}
\title{This function calculates the probability of obtain an exact amount of correct predictions}
\usage{
probabilities(line, incorrect = 0)
}
\arguments{
\item{A}{vector of 13 probabilites (line) and the nuber of incorrect predcitions, default = 0}
}
\description{
This function calculates the probability of obtain an exact amount of correct predictions
}
\examples{
cat_function()
}
\keyword{cats}

