% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odds.R
\name{to.fractional}
\alias{to.fractional}
\title{Convert Fractional odds to decimal odds.}
\usage{
to.fractional(decimal)
}
\arguments{
\item{fractional}{A character vector of fractional odds.}
}
\value{
A numeric vector of decimal odds.
}
\description{
Convert Fractional odds to decimal odds.
}
\examples{
to.fractional ("1/2")
}

